Refactored signal generators into multiple files and implemented wave table version
