+++ src/signal/generator/Noise.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace ABadCafe\Synth\Signal\Generator;
+
+use ABadCafe\Synth\Signal\ILimits;
+use ABadCafe\Synth\Signal\Context;
+use ABadCafe\Synth\Signal\Packet;
+use ABadCafe\Synth\Signal\IGenerator;
+use \SPLFixedArray;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+/**
+ * Noise - all frequencies
+ *
+ * Maps to a randomised value, irrespective of input
+ */
+class Noise extends NonFlat {
+
+    const F_PERIOD = 1.0;
+
+    protected
+        $fScaleLevel
+    ;
+
+    /**
+     * @inheritdoc
+     */
+    public function getPeriod() : float {
+        return self::F_PERIOD;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function map(Packet $oInput) : Packet {
+        $oOutput = clone $oInput;
+        $oValues = $oOutput->getValues();
+        foreach($oValues as $i => $fValue) {
+            $oValues[$i] = $this->fMinLevel + mt_rand() * $this->fScaleLevel;
+        }
+        return $oOutput;
+    }
+
+    /**
+     * @overriden
+     */
+    protected function init() {
+        $this->fScaleLevel = ($this->fMaxLevel - $this->fMinLevel) / mt_getrandmax();
+    }
+}
